+++ packages/transaction-messages/src/codecs/index.ts
@@ -1,2 +1,6 @@
+// TODO later: reduce this to only message -
+// when we've moved everything to transaction-messages
+// and don't need to export them for transactions
 export * from './address-table-lookup';
 export * from './header';
+export * from './instruction';
