+++ packages/transaction-messages/src/compile/index.ts
@@ -1,3 +1,6 @@
+// TODO later: can probably remove these exports,
+// we don't export compile-* from transactions
 export * from './accounts';
 export * from './address-table-lookups';
 export * from './header';
+export * from './instructions';
