+++ packages/transaction-messages/src/index.ts
@@ -1,6 +1,6 @@
 export * from './blockhash';
 export * from './codecs';
-export * from './compile';
+export * from './compile'; // TODO later: can probably delete this at the end
 export * from './create-transaction-message';
 export * from './durable-nonce';
 export * from './fee-payer';
